services/order-service/src/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-jose[cryptography]
python-dotenv
pydantic[email]
bcrypt==4.0.1
passlib[bcrypt]
httpx
